moved recurrent sim to simulation.py
